Set things up to treat the submodule repos as read-only.
This is ... inelegant.

Due to the smattering of files in different directories, I had some
hoops to jump through to make this work.

I also added a docker-first-run.sh script that does a lot of what the
Dockerfile used to do and invokes the db setup and card import.

The downside here is that this is realllllly clunky.
The upside is that this doesn't touch the other git repos at all.
We could consider just making variables or linking to them from this dir
instead of using submodules, but that is worth a separate discussion.